chore: move tui into cli app
