+++ apps/cli/package.json
@@ -16,12 +16,16 @@
   "dependencies": {
     "@open-harness/agent": "workspace:*",
     "@open-harness/sandbox": "workspace:*",
-    "@open-harness/tui": "workspace:*",
+    "@open-harness/shared": "workspace:*",
+    "@opentui/core": "^0.1.75",
+    "@opentui/react": "^0.1.75",
     "ink": "^6.6.0",
     "ink-spinner": "^5.0.0",
+    "marked": "^17.0.1",
     "react": "^19.2.3"
   },
   "devDependencies": {
+    "@ai-sdk/react": "^3.0.3",
     "@open-harness/tsconfig": "workspace:*",
     "@types/bun": "latest",
     "@types/node": "^22",
