+++ apps/cli/tui/chat-context.tsx
