+++ apps/cli/tui/components/approval-panel.tsx
