+++ apps/cli/tui/components/diff-view.tsx
