+++ apps/cli/tui/components/header.tsx
