+++ apps/cli/tui/components/resume-panel.tsx
