+++ apps/cli/tui/components/status-bar.tsx
