+++ apps/cli/tui/components/suggestions.tsx
