+++ apps/cli/tui/components/task-group-view.tsx
