+++ apps/cli/tui/components/text-input.tsx
