+++ apps/cli/tui/components/text-output.tsx
