+++ apps/cli/tui/components/tool-call.tsx
