+++ apps/cli/tui/components/tool-renderers/bash-renderer.tsx
