+++ apps/cli/tui/components/tool-renderers/default-renderer.tsx
