+++ apps/cli/tui/components/tool-renderers/glob-renderer.tsx
