+++ apps/cli/tui/components/tool-renderers/grep-renderer.tsx
