+++ apps/cli/tui/components/tool-renderers/index.ts
