+++ apps/cli/tui/components/tool-renderers/read-renderer.tsx
