+++ apps/cli/tui/components/tool-renderers/shared.tsx
