+++ apps/cli/tui/components/tool-renderers/write-renderer.tsx
