+++ apps/cli/tui/lib/approval.ts
