+++ apps/cli/tui/lib/colors.ts
