+++ apps/cli/tui/lib/fetch-models.ts
