+++ apps/cli/tui/lib/file-suggestions.ts
