+++ apps/cli/tui/lib/highlighter.ts
