+++ apps/cli/tui/lib/keyboard.ts
