+++ apps/cli/tui/lib/models.ts
