+++ apps/cli/tui/lib/render-tool.tsx
