+++ apps/cli/tui/lib/session-types.ts
