+++ apps/cli/tui/lib/slash-commands.ts
