+++ apps/cli/tui/lib/text-clipboard.ts
