+++ apps/cli/tui/lib/truncate.ts
