--- packages/tui/package.json
@@ -1,35 +1,0 @@
-{
-  "name": "@open-harness/tui",
-  "version": "0.0.0",
-  "private": true,
-  "type": "module",
-  "main": "./index.tsx",
-  "exports": {
-    ".": "./index.tsx"
-  },
-  "scripts": {
-    "typecheck": "tsc --noEmit",
-    "lint": "oxlint",
-    "lint:fix": "oxlint --fix"
-  },
-  "dependencies": {
-    "@open-harness/agent": "workspace:*",
-    "@open-harness/sandbox": "workspace:*",
-    "@open-harness/shared": "workspace:*",
-    "@opentui/core": "^0.1.75",
-    "@opentui/react": "^0.1.75",
-    "marked": "^17.0.1",
-    "react": "^19.2.3"
-  },
-  "devDependencies": {
-    "@ai-sdk/react": "^3.0.3",
-    "@open-harness/tsconfig": "workspace:*",
-    "@types/bun": "latest",
-    "@types/node": "^22",
-    "@types/react": "^19.2.7",
-    "oxlint": "^1.38.0"
-  },
-  "peerDependencies": {
-    "typescript": "^5"
-  }
-}
